Keep left-side indentation
As the comment suggests, the strip is done to remove any trailing
white-space so the dumper does not get confused. It's also mentioned
that the trailing white-space is irrelevant for the scripts.

However, the leading white-space might be significant - python scripts,
bash here-documents, etc. In any case it at least improves the
readability.